# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ticketopia API
  version: 1.5.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Read-only event discovery across all Ticketmaster categories, shared with the
    web UI. Event, artist, and venue IDs are provider-scoped (ticketmaster:SOURCE_ID).
    Category and genre IDs are Ticketmaster
    taxonomy IDs. Data is cached, not a live inventory or durable event history.
    Opt-in Google authentication provides MariaDB-backed profiles and private
    preferences, private durable event saves, independent event interest, and public event recommendations.
    Interest identity is private by default; counts include private choices but
    participant lists and profile activity include only explicitly public choices.
    Account APIs use a browser session or a user-scoped API token;
    discovery remains public. Account APIs return 503 when accounts are disabled.
servers:
  - url: /
security: []
paths:
...
  /api/v1/events:
    get:
      operationId: listEvents
      summary: Search events across categories
      description: |
        Dates are inclusive venue-local dates, defaulting to today in UTC through
        90 days later. Maximum span is 366 days. Results use Ticketmaster's
        date/name ordering. Pass the same explicit filters with the next cursor.
        Provider pages can shift on refresh; cursors do not pin a snapshot.
        Unknown filters, repeated parameters, and invalid cursors return 400.
        Searches default to all categories. Legacy genre-only searches imply Music.
        Cursors issued before category support must be replaced by a new search.
      parameters:
        - {name: city, in: query, schema: {type: string, maxLength: 120}}
        - {name: country, in: query, description: Two-letter country code., schema: {type: string, pattern: '^[A-Za-z]{2}$'}}
        - {name: keyword, in: query, schema: {type: string, maxLength: 120}}
        - name: category_id
          in: query
          description: 'Ticketmaster segment ID from /api/v1/categories, or all. Omitted/empty means all categories unless genre_id is supplied, which implies Music for legacy links.'
          schema: {type: string, pattern: '^[A-Za-z0-9_-]{1,128}$'}
        - {name: genre_id, in: query, description: 'ID from the selected category’s genres. Explicit category_id=all cannot be combined with a genre. Category/genre combinations follow provider matching and can return no results.', schema: {type: string}}
        - {name: artist_id, in: query, schema: {$ref: '#/components/schemas/ResourceID'}}
        - {name: venue_id, in: query, schema: {$ref: '#/components/schemas/ResourceID'}}
        - {name: start_date, in: query, schema: {type: string, format: date}}
        - {name: end_date, in: query, schema: {type: string, format: date}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 512}}
      responses:
        '200':
          description: Events and source freshness. Empty results are cached too.
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor, total, limited, meta]
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/Event'}}
                  next_cursor: {type: [string, 'null']}
                  total: {type: integer, minimum: 0, description: 'Total reported by Ticketmaster, including results beyond its paging limit.'}
                  limited: {type: boolean, description: True when further pages exist but the provider's size times page limit prevents access. Narrow the filters.}
                  meta: {$ref: '#/components/schemas/Freshness'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/me/event-interests:
    get:
      operationId: listOwnEventInterests
      summary: List your event interest, including private choices
      description: |
        Last-known event snapshots without provider calls, ordered by interested_at
        then binary event ID descending. Cursors are owner- and audience-bound.
        New choices appear on refresh; removals and visibility changes can shorten
        later pages. Unknown/repeated parameters are rejected. Responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/InterestCursor'}
      responses:
        '200':
          description: Your event interest, or an empty array.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventInterestList'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
  /api/v1/me/event-interests/states:
    get:
      operationId: getOwnEventInterestStates
      summary: Read aggregate counts and your own choices in a bounded batch
      description: |
        Repeat event_id up to 101 times; no other parameters or provider calls.
        Counts include private and public choices. Interested/visibility belong
        only to the authenticated owner; no other account identities are returned.
        Browser responses include csrf_token; bearer responses omit it.
        default_visibility applies to new choices only. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - name: event_id
          in: query
          style: form
          explode: true
          schema:
            type: array
            maxItems: 101
            items: {$ref: '#/components/schemas/ResourceID'}
      responses:
        '200':
          description: A state for each requested ID; unknown IDs have count zero and interested false.
          content:
            application/json:
              schema:
                type: object
                required: [states, default_visibility]
                properties:
                  states: {type: object, additionalProperties: {$ref: '#/components/schemas/EventInterestState'}}
                  csrf_token: {type: string}
                  default_visibility: {type: string, enum: [private, public]}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
  /api/v1/me/event-interests/{event_id}:
    parameters:
      - {name: event_id, in: path, required: true, schema: {$ref: '#/components/schemas/ResourceID'}}
    get:
      operationId: getOwnEventInterest
      summary: Read your interest in an event
      description: Owner-only. No query parameters. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      responses:
        '200':
          description: Your choice and retained event metadata.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventInterest'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/InterestNotFound'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
    put:
      operationId: setOwnEventInterest
      summary: Mark Interested or change the visibility of your choice
      description: |
        A JSON object is required; send {} or {"visibility":"private"} or
        {"visibility":"public"}. Omission preserves an existing choice, or uses
        the account's private-by-default profile setting for a new choice.
        Public choices disclose your public profile on event participant lists
        and the event on your public profile. Changing the profile default never
        changes existing choices. No query parameters or client-supplied owner or metadata.
        Identity, snapshot, and first interest commit atomically. Repeated PUT
        preserves interested_at and needs no provider call for an existing choice.
        Separate from saves, recommendations, tickets, and attendance.
        Cookie writes require CSRF and same-origin checks; bearer writes do not.
        All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                visibility: {type: string, enum: [private, public]}
      responses:
        '201':
          description: Interest created.
          headers:
            Location: {description: Owner-only interest resource., schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventInterest'}
        '200':
          description: Existing interest, retaining its original interested_at.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventInterest'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFRejected'}
        '404': {$ref: '#/components/responses/InterestNotFound'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
    delete:
      operationId: removeOwnEventInterest
      summary: Remove your event interest without changing your save
      description: Empty body and no query parameters. Owner-only and idempotent, including when absent. Retains shared event metadata and other users' interest. Cookie writes require CSRF and same-origin checks. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      responses:
        '204': {description: Interest removed or already absent.}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFRejected'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
  /api/v1/events/{event_id}/interest:
    get:
      operationId: getEventInterestCount
      summary: Read an event's aggregate interest count without identities
      description: Public read including private choices. No query parameters or provider calls. Unknown event IDs have count zero. All responses are private, no-store, so visibility changes take effect on the next read.
      parameters:
        - {name: event_id, in: path, required: true, schema: {$ref: '#/components/schemas/ResourceID'}}
      responses:
        '200':
          description: Aggregate only; no personal state or identity.
          content:
            application/json:
              schema:
                type: object
                required: [count]
                additionalProperties: false
                properties:
                  count: {type: integer, minimum: 0}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
  /api/v1/events/{event_id}/interested-users:
    get:
      operationId: listPublicInterestedUsers
      summary: List only explicitly public interest identities for an event
      description: |
        Public allowlisted profiles only, never email or private choices. Ordered
        by interested_at then binary account ID descending. Cursors are bound to
        this event. No provider calls; unknown events return an empty array.
        Removals/visibility changes can shorten pages. All responses are private, no-store.
      parameters:
        - {name: event_id, in: path, required: true, schema: {$ref: '#/components/schemas/ResourceID'}}
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/InterestCursor'}
      responses:
        '200':
          description: Public profiles with their interest timestamps, or an empty array.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterestParticipants'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
  /api/v1/users/{user_id}/event-interests:
    get:
      operationId: listPublicProfileEventInterests
      summary: List an account's explicitly public event interest
      description: |
        Public event activity only; private choices and saves are never included,
        even for the owner. Last-known snapshots require no provider calls. Ordered
        by interested_at then binary event ID descending. Cursors are account-
        and audience-bound. All responses are private, no-store.
      parameters:
        - {name: user_id, in: path, required: true, schema: {$ref: '#/components/schemas/AccountID'}}
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/InterestCursor'}
      responses:
        '200':
          description: Only public interest, with visibility always public.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventInterestList'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '404':
          description: Account not found, including an invalid account ID.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/Problem'}
        '503': {$ref: '#/components/responses/InterestsUnavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ticketopia_session
      description: 'HttpOnly, Secure, SameSite=Lax cookie issued after Google sign-in. Local HTTP development uses ticketopia_session instead.'
    bearerToken:
      type: http
      scheme: bearer
      description: Opaque, expiring user-scoped API credential created in an authenticated browser. Not a Google access token or ID token.
  parameters:
...
    InterestLimit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    InterestCursor:
      name: cursor
      in: query
      description: Opaque next_cursor from this collection and audience.
      schema: {type: string, maxLength: 512}
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: Required for every cookie-authenticated write. Read csrf_token from GET /api/v1/me. Not required with an API bearer token. Supplied Origin/Referer must match AUTH_BASE_URL.
      schema: {type: string}
  responses:
...
    InterestNotFound:
      description: Event cannot be resolved for a new choice, or this owner has no interest in it.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    InterestsUnavailable:
      description: Accounts or interest are disabled, durable storage is unavailable, or a new event cannot be resolved and no snapshot exists. No in-memory fallback.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    AccountBadRequest:
      description: Unsupported query, invalid JSON/form fields or values, or a request body exceeding 16 KiB.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthenticated:
      description: Missing, expired, revoked, wrong-kind, or invalid credentials. Credential management requires a browser session.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    CSRFRejected:
      description: Invalid or missing CSRF token, or a cross-origin cookie write.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    BadRequest:
      description: Invalid query or ID, or filters rejected by Ticketmaster. Local validation failures make no provider request.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unavailable:
      description: Provider outage, timeout, missing key, or exhausted request budget with no usable cached data.
      headers:
        Retry-After:
          description: Seconds until retry, when known.
          schema: {type: integer, minimum: 1}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    EventInterest:
      type: object
      required: [event, interested_at, visibility, meta]
      properties:
        event: {$ref: '#/components/schemas/Event'}
        interested_at: {type: string, format: date-time, description: 'Original choice time, preserved by retries and visibility changes.'}
        visibility: {type: string, enum: [private, public]}
        meta: {$ref: '#/components/schemas/Freshness'}
    EventInterestList:
      type: object
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/EventInterest'}}
        next_cursor: {type: [string, 'null']}
    EventInterestState:
      type: object
      required: [count, interested]
      additionalProperties: false
      properties:
        count: {type: integer, minimum: 0}
        interested: {type: boolean}
        visibility: {type: string, enum: [private, public], description: Present only when this owner is interested.}
    InterestParticipants:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            type: object
            required: [profile, interested_at]
            properties:
              profile: {$ref: '#/components/schemas/PublicProfile'}
              interested_at: {type: string, format: date-time}
        next_cursor: {type: [string, 'null']}
...
    AccountID:
      type: string
      pattern: '^[0-9a-f]{32}$'
    PublicProfile:
      type: object
      additionalProperties: false
      required: [id, display_name, bio]
      properties:
        id: {$ref: '#/components/schemas/AccountID'}
        display_name: {type: string, minLength: 1, maxLength: 80}
        bio: {type: string, maxLength: 500}
...
    ResourceID:
      type: string
      pattern: '^ticketmaster:[A-Za-z0-9_-]{1,128}$'
    Freshness:
      type: object
      required: [data_as_of, stale]
      properties:
        data_as_of: {type: string, format: date-time, description: 'When this data was successfully collected, not when this response was served.'}
        stale: {type: boolean, description: A refresh failed or was paused and retained data is being served.}
    Source:
      type: object
      required: [provider, id, url]
      properties:
        provider: {type: string, const: ticketmaster}
        id: {type: string}
        url: {type: string, description: 'Original source URL, or an empty string if unavailable.'}
    NamedID:
      type: object
      required: [id, name]
      properties:
        id: {type: string}
        name: {type: string}
...
    NullableInstant:
      type: [string, 'null']
      format: date-time
    EventStart:
      type: object
      required: [date_time, local_date, local_time, timezone, date_tba, date_tbd, time_tba, no_specific_time]
      properties:
        date_time: {$ref: '#/components/schemas/NullableInstant'}
        local_date: {type: [string, 'null'], description: Provider local date; TBA/TBD flags override any placeholder value.}
        local_time: {type: [string, 'null'], description: Provider local time; time_tba and no_specific_time override placeholder values.}
        timezone: {type: [string, 'null']}
        date_tba: {type: boolean}
        date_tbd: {type: boolean}
        time_tba: {type: boolean}
        no_specific_time: {type: boolean}
    PublicSale:
      type: object
      required: [start, end, start_tbd]
      properties:
        start: {$ref: '#/components/schemas/NullableInstant'}
        end: {$ref: '#/components/schemas/NullableInstant'}
        start_tbd: {type: boolean}
    Presale:
      type: object
      required: [name, description, url, start, end]
      properties:
        name: {type: string}
        description: {type: string}
        url: {type: string}
        start: {$ref: '#/components/schemas/NullableInstant'}
        end: {$ref: '#/components/schemas/NullableInstant'}
    PriceRange:
      type: object
      required: [type, currency, min, max, fees_included]
      properties:
        type: {type: string}
        currency: {type: string}
        min: {type: [string, 'null'], description: Decimal amount preserved without floating-point conversion.}
        max: {type: [string, 'null'], description: Decimal amount preserved without floating-point conversion.}
        fees_included: {type: [boolean, 'null'], description: Null when the source does not establish fee inclusion.}
    Image:
      type: object
      required: [url, width, height, ratio, fallback]
      properties:
        url: {type: string}
        width: {type: integer}
        height: {type: integer}
        ratio: {type: string}
        fallback: {type: boolean}
    Artist:
      type: object
      required: [id, name, source]
      properties:
        id: {$ref: '#/components/schemas/ResourceID'}
        name: {type: string}
        source: {$ref: '#/components/schemas/Source'}
    Place:
      type: object
      required: [name, address, city, state, country, country_code, postal_code, latitude, longitude]
      properties:
        name: {type: string}
        address: {type: string}
        city: {type: string}
        state: {type: string}
        country: {type: string}
        country_code: {type: string}
        postal_code: {type: string}
        latitude: {type: string}
        longitude: {type: string}
    Venue:
      allOf:
        - {$ref: '#/components/schemas/Place'}
        - type: object
          required: [id, source, timezone]
          properties:
            id: {$ref: '#/components/schemas/ResourceID'}
            source: {$ref: '#/components/schemas/Source'}
            timezone: {type: string}
    Classification:
      type: object
      required: [primary, segment, genre, subgenre]
      properties:
        primary: {type: boolean}
        segment: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
        genre: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
        subgenre: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
    Event:
      type: object
      required: [id, name, source, start, status, public_sale, presales, price_ranges, images, artists, venues, place, classifications, info, please_note]
      properties:
        id: {$ref: '#/components/schemas/ResourceID'}
        name: {type: string}
        source: {$ref: '#/components/schemas/Source'}
        start: {$ref: '#/components/schemas/EventStart'}
        status: {type: string, description: 'Source status (onsale, offsale, canceled, postponed, rescheduled) or unknown. Not seat availability.'}
        public_sale: {$ref: '#/components/schemas/PublicSale'}
        presales: {type: array, items: {$ref: '#/components/schemas/Presale'}}
        price_ranges: {type: [array, 'null'], items: {$ref: '#/components/schemas/PriceRange'}, description: 'Null means no advertised prices, not free admission.'}
        images: {type: array, items: {$ref: '#/components/schemas/Image'}}
        artists: {type: array, items: {$ref: '#/components/schemas/Artist'}, description: 'Provider attractions, including performers or teams. The field name is retained for compatibility; events without attractions return an empty array.'}
        venues: {type: array, items: {$ref: '#/components/schemas/Venue'}}
        place: {anyOf: [{$ref: '#/components/schemas/Place'}, {type: 'null'}]}
        classifications: {type: array, items: {$ref: '#/components/schemas/Classification'}}
        info: {type: string}
        please_note: {type: string}
    Problem:
      type: object
      required: [type, title, status, code, detail]
      properties:
        type: {type: string, const: 'about:blank'}
        title: {type: string}
        status: {type: integer}
        code: {type: string, enum: [invalid_query, event_not_found, provider_unavailable, invalid_input, unauthenticated, csrf_failed, invalid_sign_in, account_not_found, rate_limited, credential_limit, accounts_disabled, accounts_unavailable]}
        detail: {type: string}
        fields: {type: object, additionalProperties: {type: string}}